requirements.txt
streamlit
tabulate
groq
python-dotenv
pandas
PyPDF2